+++ src/textual/_layout.py
@@ -137,9 +137,4 @@ class Layout(ABC):
             arrangement = widget._arrange(Size(width, 0))
             height = arrangement.total_region.bottom
 
-            if widget.styles.max_height and any(
-                widget.styles.is_relative_height for widget in widget.children
-            ):
-                height = max(height, container.height)
-
         return height