+++ src/textual/layouts/horizontal.py
@@ -23,10 +23,23 @@ class HorizontalLayout(Layout):
         x = max_height = Fraction(0)
         parent_size = parent.outer_size
 
+        child_styles = [child.styles for child in children]
+        box_margins = [styles.margin for styles in child_styles]
+        if box_margins:
+            total_margin = sum(
+                max(margin1.right, margin2.left)
+                for margin1, margin2 in zip(box_margins, box_margins[1:])
+            )
+            if len(box_margins) > 1:
+                total_margin += box_margins[0].left + box_margins[-1].right
+        else:
+            total_margin = 0
+
         box_models = resolve_box_models(
             [child.styles.width for child in children],
             children,
             size,
+            size - Size(total_margin, 0),
             parent_size,
             dimension="width",
         )