+++ src/textual/layouts/vertical.py
@@ -23,10 +23,24 @@ class VerticalLayout(Layout):
         add_placement = placements.append
         parent_size = parent.outer_size
 
+        child_styles = [child.styles for child in children]
+        box_margins = [styles.margin for styles in child_styles]
+        if box_margins:
+            total_margin = sum(
+                max(margin1.bottom, margin2.top)
+                for margin1, margin2 in zip(box_margins, box_margins[1:])
+            )
+            if len(box_margins) > 1:
+                total_margin += box_margins[0].top + box_margins[-1].bottom
+
+        else:
+            total_margin = 0
+
         box_models = resolve_box_models(
-            [child.styles.height for child in children],
+            [styles.height for styles in child_styles],
             children,
             size,
+            size - Size(0, total_margin),
             parent_size,
             dimension="height",
         )