+++ tests/snapshot_tests/test_snapshots.py
@@ -386,3 +386,7 @@ def test_modal_dialog_bindings(snap_compare):
 def test_dock_scroll(snap_compare):
     # https://github.com/Textualize/textual/issues/2188
     assert snap_compare(SNAPSHOT_APPS_DIR / "dock_scroll.py", terminal_size=(80, 25))
+
+
+def test_auto_fr(snap_compare):
+    assert snap_compare(SNAPSHOT_APPS_DIR / "auto_fr.py", terminal_size=(80, 25))