+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Allowed border_title and border_subtitle to accept Text objects
 - Added additional line around titles
 
+### Fixed
+
+- Fixed margin not being respected when width or height is "auto"
+- When a container is auto, relative dimensions in children stretch the container.
+
 ## [0.18.0] - 2023-04-04
 
 ### Added